backend/services/dirty_floor_service/requirements.txt
requests
redis
torch
torchvision
transformers
python-dotenv
numpy
Pillow
opencv-python
peft
hqq
accelerate
